remove character stuff
