+++ packages/agent/src/server/api/api-utils.ts
@@ -1,6 +1,5 @@
-import { type UUID, validateUuid } from '@elizaos/core';
+import { logger, type UUID } from '@elizaos/core';
 import type { Response } from 'express';
-import { logger } from '@elizaos/core';
 
 export interface UUIDParams {
     agentId: UUID;
@@ -33,3 +32,31 @@ export function handleValidationError(error: unknown, res: Response, context: st
         });
     }
 } 
+
+
+import type { Character } from '@elizaos/core';
+import { validateCharacterConfig } from '@elizaos/core';
+import { createDatabaseAdapter } from '@elizaos/plugin-sql';
+import type { Request } from 'express';
+
+// Create database adapter from environment variables.
+const dbAdapter = createDatabaseAdapter({
+    dataDir: process.env.PGLITE_DATA_DIR,
+    postgresUrl: process.env.POSTGRES_URL,
+});
+
+// Utility function to parse and validate character data.
+function getValidatedCharacter(req: Request): Character {
+    return validateCharacterConfig(req.body) as Character;
+}
+
+// Utility function to check if a character exists.
+async function ensureCharacterExists(name: string, res: Response): Promise<Character | null> {
+    const character = await dbAdapter.getCharacter(name);
+    if (!character) {
+        res.status(404).json({ error: 'Character not found' });
+        return null;
+    }
+    return character;
+}
+
