+++ packages/agent/src/server/api/index.ts
@@ -1,12 +1,11 @@
-import express from 'express';
+import type { IAgentRuntime } from '@elizaos/core';
+import { getEnvVariable } from '@elizaos/core';
 import bodyParser from 'body-parser';
 import cors from 'cors';
-import { getEnvVariable } from '@elizaos/core';
-import { agentRouter } from './agent.ts';
-import { characterRouter } from './character.ts';
-import { teeRouter } from './tee.ts';
+import express from 'express';
 import type { AgentServer } from '..';
-import type { IAgentRuntime } from '@elizaos/core';
+import { agentRouter } from './agent.ts';
+import { teeRouter } from './tee.ts';
 
 export function createApiRouter(
     agents: Map<string, IAgentRuntime>,
@@ -35,7 +34,6 @@ export function createApiRouter(
 
     // Mount sub-routers
     router.use('/agents', agentRouter(agents, directClient));
-    router.use('/characters', characterRouter());
     router.use('/tee', teeRouter(agents));
 
     // Health check endpoints
