+++ packages/client/src/App.tsx
@@ -8,13 +8,10 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Chat from "./routes/chat";
 import Overview from "./routes/overview";
 import Home from "./routes/home";
-import Characters from "./routes/characters";
-import NewCharacter from "./routes/new-character";
 import useVersion from "./hooks/use-version";
 import { useEffect } from "react";
 import { apiClient } from "./lib/api";
 import { STALE_TIMES } from "./hooks/use-query-hooks";
-import EditCharacter from "./routes/edit-character";
 
 // Create a query client with optimized settings
 const queryClient = new QueryClient({
@@ -43,18 +40,11 @@ const queryClient = new QueryClient({
 // Prefetch initial data with smarter error handling
 const prefetchInitialData = async () => {
   try {
-    // Prefetch characters
-    await queryClient.prefetchQuery({
-      queryKey: ["characters"],
-      queryFn: () => apiClient.getCharacters(),
-      staleTime: STALE_TIMES.STANDARD,
-    });
-    
     // Prefetch agents (real-time data so shorter stale time)
     await queryClient.prefetchQuery({
       queryKey: ["agents"],
       queryFn: () => apiClient.getAgents(),
-      staleTime: STALE_TIMES.FREQUENT, 
+      staleTime: STALE_TIMES.FREQUENT,
     });
   } catch (error) {
     console.error("Error prefetching initial data:", error);
@@ -89,9 +79,6 @@ function App() {
                                 <div className="flex flex-1 flex-col gap-4 size-full container">
                                     <Routes>
                                         <Route path="/" element={<Home />} />
-                                        <Route path="/characters" element={<Characters />} />
-                                        <Route path="/characters/new" element={<NewCharacter />} />
-                                        <Route path="/characters/edit/:characterName" element={<EditCharacter />} />
                                         <Route
                                             path="chat/:agentId"
                                             element={<Chat />}
