+++ packages/client/src/hooks/use-query-hooks.ts
@@ -128,41 +128,6 @@ export function useAgent(agentId: UUID | undefined | null, options = {}) {
   });
 }
 
-// Hook for fetching characters with smart polling (less frequent)
-export function useCharacters(options = {}) {
-  const network = useNetworkStatus();
-  
-  return useQuery({
-    queryKey: ['characters'],
-    queryFn: () => apiClient.getCharacters(),
-    staleTime: STALE_TIMES.STANDARD,
-    // Characters change less frequently, so poll less often
-    refetchInterval: !network.isOffline 
-      ? STALE_TIMES.RARE 
-      : false,
-    refetchIntervalInBackground: false,
-    ...options
-  });
-}
-
-// Hook for fetching a specific character with smart polling
-export function useCharacter(characterName: string | undefined | null, isUrlEncoded = false, options = {}) {
-  const network = useNetworkStatus();
-  
-  return useQuery({
-    queryKey: ['character', characterName],
-    queryFn: () => apiClient.getCharacter(characterName || '', isUrlEncoded),
-    staleTime: STALE_TIMES.STANDARD,
-    enabled: Boolean(characterName),
-    // Characters change less frequently, so poll less often
-    refetchInterval: !network.isOffline 
-      ? STALE_TIMES.RARE 
-      : false,
-    refetchIntervalInBackground: false,
-    ...options
-  });
-}
-
 // Hook for starting an agent with optimistic updates
 export function useStartAgent() {
   const queryClient = useQueryClient();
@@ -235,123 +200,8 @@ export function useStopAgent() {
   });
 }
 
-// Hook for creating a character with optimistic updates
-export function useCreateCharacter() {
-  const queryClient = useQueryClient();
-  const { toast } = useToast();
-  
-  return useMutation({
-    mutationFn: (character: Character) => apiClient.createCharacter(character),
-    onSuccess: (data) => {
-      // Optimistically update characters list
-      queryClient.invalidateQueries({ queryKey: ['characters'] });
-      
-      toast({
-        title: 'Character Created',
-        description: `Successfully created character: ${data?.name || 'Unknown'}`,
-      });
-    },
-    onError: (error) => {
-      toast({
-        title: 'Error',
-        description: error instanceof Error ? error.message : 'Failed to create character',
-        variant: 'destructive',
-      });
-    }
-  });
-}
 
-// Hook for updating a character with optimistic updates
-export function useUpdateCharacter() {
-  const queryClient = useQueryClient();
-  const { toast } = useToast();
-  
-  return useMutation({
-    mutationFn: ({ characterName, character, isUrlEncoded = false }: { 
-      characterName: string;
-      character: Character;
-      isUrlEncoded?: boolean;
-    }) => apiClient.updateCharacter(characterName, character, isUrlEncoded),
-    onSuccess: (_, variables) => {
-      // Optimistically update character data
-      queryClient.invalidateQueries({ queryKey: ['character', variables.characterName] });
-      queryClient.invalidateQueries({ queryKey: ['characters'] });
-      
-      toast({
-        title: 'Character Updated',
-        description: `Successfully updated character: ${variables.character.name}`,
-      });
-    },
-    onError: (error) => {
-      toast({
-        title: 'Error',
-        description: error instanceof Error ? error.message : 'Failed to update character',
-        variant: 'destructive',
-      });
-    }
-  });
-}
 
-// Define the structure of character data response
-interface CharactersResponse {
-  characters: Character[];
-  [key: string]: unknown;
-}
-
-// Hook for deleting a character with optimistic updates
-export function useDeleteCharacter() {
-  const queryClient = useQueryClient();
-  const { toast } = useToast();
-  
-  return useMutation({
-    mutationFn: ({ characterName, isUrlEncoded = false }: { 
-      characterName: string;
-      isUrlEncoded?: boolean;
-    }) => apiClient.removeCharacter(characterName, isUrlEncoded),
-    onMutate: async (variables) => {
-      // Optimistically remove from the characters list
-      const previousCharacters = queryClient.getQueryData<CharactersResponse>(['characters']);
-      
-      queryClient.setQueryData<CharactersResponse>(['characters'], (old) => {
-        if (!old || !old.characters) return old;
-        
-        return {
-          ...old,
-          characters: old.characters.filter(
-            (c: Character) => c.name !== variables.characterName
-          )
-        };
-      });
-      
-      return { previousCharacters };
-    },
-    onSuccess: (_, variables) => {
-      // Remove from the cache
-      queryClient.removeQueries({ queryKey: ['character', variables.characterName] });
-      
-      toast({
-        title: 'Character Deleted',
-        description: `Successfully deleted character: ${variables.characterName}`,
-      });
-    },
-    onError: (error, _, context) => {
-      // Restore the previous data
-      if (context?.previousCharacters) {
-        queryClient.setQueryData(['characters'], context.previousCharacters);
-      }
-      
-      toast({
-        title: 'Error',
-        description: error instanceof Error ? error.message : 'Failed to delete character',
-        variant: 'destructive',
-      });
-    },
-    onSettled: () => {
-      // Always refetch to ensure consistency
-      queryClient.invalidateQueries({ queryKey: ['characters'] });
-    }
-  });
-}
 
 // Hook for fetching messages directly for a specific agent without requiring a room
 export function useAgentMessages(agentId: UUID) {
