+++ packages/client/tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/types.ts","./src/vite-env.d.ts","./src/components/app-sidebar.tsx","./src/components/array-input.tsx","./src/components/audio-recorder.tsx","./src/components/chat.tsx","./src/components/connection-status.tsx","./src/components/copy-button.tsx","./src/components/input-copy.tsx","./src/components/overview.tsx","./src/components/page-title.tsx","./src/components/ui/avatar.tsx","./src/components/ui/badge.tsx","./src/components/ui/breadcrumb.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/collapsible.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/separator.tsx","./src/components/ui/sheet.tsx","./src/components/ui/sidebar.tsx","./src/components/ui/skeleton.tsx","./src/components/ui/tabs.tsx","./src/components/ui/textarea.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/components/ui/tooltip.tsx","./src/components/ui/chat/chat-bubble.tsx","./src/components/ui/chat/chat-input.tsx","./src/components/ui/chat/chat-message-list.tsx","./src/components/ui/chat/chat-tts-button.tsx","./src/components/ui/chat/expandable-chat.tsx","./src/components/ui/chat/message-loading.tsx","./src/components/ui/chat/hooks/useautoscroll.tsx","./src/hooks/use-mobile.tsx","./src/hooks/use-toast.ts","./src/hooks/use-version.tsx","./src/lib/api.ts","./src/lib/utils.ts","./src/routes/chat.tsx","./src/routes/home.tsx","./src/routes/overview.tsx","./src/types/index.ts"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/types.ts","./src/vite-env.d.ts","./src/components/app-sidebar.tsx","./src/components/array-input.tsx","./src/components/audio-recorder.tsx","./src/components/character-form.tsx","./src/components/chat.tsx","./src/components/connection-status.tsx","./src/components/copy-button.tsx","./src/components/input-copy.tsx","./src/components/overview.tsx","./src/components/page-title.tsx","./src/components/ui/action-card.tsx","./src/components/ui/avatar.tsx","./src/components/ui/badge.tsx","./src/components/ui/button.tsx","./src/components/ui/card-actions.tsx","./src/components/ui/card.tsx","./src/components/ui/collapsible.tsx","./src/components/ui/command.tsx","./src/components/ui/dialog.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/separator.tsx","./src/components/ui/sheet.tsx","./src/components/ui/sidebar.tsx","./src/components/ui/skeleton.tsx","./src/components/ui/tabs.tsx","./src/components/ui/textarea.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/components/ui/tooltip.tsx","./src/components/ui/chat/chat-bubble.tsx","./src/components/ui/chat/chat-input.tsx","./src/components/ui/chat/chat-message-list.tsx","./src/components/ui/chat/chat-tts-button.tsx","./src/components/ui/chat/expandable-chat.tsx","./src/components/ui/chat/message-loading.tsx","./src/components/ui/chat/hooks/useautoscroll.tsx","./src/hooks/use-mobile.tsx","./src/hooks/use-plugins.ts","./src/hooks/use-query-hooks.ts","./src/hooks/use-toast.ts","./src/hooks/use-version.tsx","./src/lib/api.ts","./src/lib/utils.ts","./src/lib/world-manager.ts","./src/routes/chat.tsx","./src/routes/home.tsx","./src/routes/overview.tsx","./src/types/index.ts"],"version":"5.6.3"}
